# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Simple Payment API",
        "contact": {}
    },
    "paths": {
...
        "/api/customers": {
            "get": {
                "tags": [
                    "customer"
                ],
                "summary": "Get all customers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.CustomerResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.Customer"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "customer"
                ],
                "summary": "Create new customer",
                "parameters": [
                    {
                        "description": "Customer",
                        "name": "string",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CustomerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CustomerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/customers/id": {
            "put": {
                "tags": [
                    "customer"
                ],
                "summary": "Topup customer by ID",
                "parameters": [
                    {
                        "description": "TopUp",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.TopUpRequest"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "balance": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Customer"
                        }
                    }
                }
            }
        },
        "/api/customers/{id}": {
            "get": {
                "tags": [
                    "customer"
                ],
                "summary": "Get customer by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Customer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Customer"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "customer"
                ],
                "summary": "Delete customer by ID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Customer"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Customer": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "customer_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "model.CustomerRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "user_id": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
        "model.CustomerResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.TopUpRequest": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer",
                    "example": 50000
                },
                "customer_id": {
                    "type": "string",
                    "example": "1"
                }
            }
        }
    }
}